Move user type to shared
