+++ server/src/features/game-popularity/updateGamePopularity.ts
@@ -3,7 +3,7 @@ import { logger } from 'server/utils/logger';
 import { updateWithSignups } from 'server/features/game-popularity/utils/updateWithSignups';
 import { updateWithAssign } from 'server/features/game-popularity/utils/updateWithAssign';
 import { config } from 'server/config';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 import { Game } from 'shared/typings/models/game';
 import { findUsers } from 'server/features/user/userRepository';
 import { findGames } from 'server/features/game/gameRepository';
