+++ server/src/features/player-assignment/munkres/utils/getSignupMatrix.ts
@@ -1,6 +1,6 @@
 import { logger } from 'server/utils/logger';
 import { Game } from 'shared/typings/models/game';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 
 export const getSignupMatrix = (
   signedGames: readonly Game[],
