+++ server/src/features/player-assignment/padg/padgAssignPlayers.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 import { Game } from 'shared/typings/models/game';
 import { PlayerAssignmentResult } from 'server/typings/result.typings';
 import { getStartingGames } from 'server/features/player-assignment/utils/getStartingGames';
