+++ server/src/features/statistics/fixer-helpers/dataAnonymizer.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import faker from 'faker';
 import { logger } from 'server/utils/logger';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 import { ResultsCollectionEntry } from 'server/typings/result.typings';
 import { writeJson } from 'server/features/statistics/statsUtil';
