+++ server/src/features/statistics/statistics-helpers/gameDataHelpers.ts
@@ -2,7 +2,7 @@ import moment from 'moment';
 import _ from 'lodash';
 import { logger } from 'server/utils/logger';
 import { Game } from 'shared/typings/models/game';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 import { getMaximumNumberOfPlayersByTime } from './resultDataHelpers';
 import {
   StringNumberObject,
