+++ server/src/features/statistics/statistics-helpers/userDataHelpers.ts
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { logger } from 'server/utils/logger';
-import { User } from 'server/typings/user.typings';
+import { User } from 'shared/typings/models/user';
 import { StringNumberObject } from 'server/typings/common.typings';
 import { toPercent } from 'server/features/statistics/statsUtil';
